// web/src/shared/api/generated/model/getAvatarConversationsParams.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Elmanhg.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { AvatarEntryPoint } from './avatarEntryPoint';

export type GetAvatarConversationsParams = {
search?: string;
entryPoint?: AvatarEntryPoint;
from?: string;
to?: string;
/**
 * @pattern ^-?(?:0|[1-9]\d*)$
 */
pageNumber?: number | string;
/**
 * @pattern ^-?(?:0|[1-9]\d*)$
 */
pageSize?: number | string;
};